+++ src/steps/privateca/constants.ts
@@ -1,5 +1,7 @@
 export const STEP_PRIVATE_CA_CERTIFICATE_AUTHORITIES =
   'fetch-private-ca-certificate-authorities';
+export const STEP_CREATE_PRIVATE_CA_CERTIFICATE_AUTHORITY_BUCKET_RELATIONSHIPS =
+  'build-private-ca-certificate-authority-bucket-relationships';
 export const STEP_PRIVATE_CA_CERTIFICATES = 'fetch-private-ca-certificates';
 
 export const ENTITY_TYPE_PRIVATE_CA_CERTIFICATE_AUTHORITY =